# orders-microservice/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: YumYumNow Orders API
  version: 1.0.0
  description: API for handling orders in the YumYumNow application.
servers:
  - url: 'https://localhost:8082'
...
paths:
...
  /customer/{customerId}/order:
    post:
      tags:
        - Customer
      summary: Create a new order
      description:  Creates a new order for the customer with a specified vendor. This is done by selecting a vendor from the list of all vendors basically, so we are sure that the customer is in the range of the vendor.
      operationId: createOrder
      parameters:
        - in: path
          name: customerId
          required: true
          schema:
            type: string
            format: uuid
          examples:
            customerIdExample:
              value: "123e4567-e89b-12d3-a456-426614174000"  
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                vendorId:
                  type: string
                  format: uuid
                  description: The ID of the restaurant the customer is ordering from.
                address:
                  $ref: '#/components/schemas/Address'
                 
            examples:
              createOrderExample:
                summary: Example of creating a new order
                value:
                  vendorId: 42
                  address:
                    houseNumber: 123
                    zip: "2133DC"
                    longitude: 40.7128
                    latitude: -74.0060      
      responses:
        '200':
          description: Newly created order object with orderId, customerId, vendorId and address populated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
              examples:
                orderResponseExample:
                  summary: Example response for a newly created order
                  value:
                    ID: "987e4567-e89b-12d3-a456-426614174002"
                    dishes: []
                    location:
                      houseNumber: 123
                      zip: "2133DC"
                      longitude: 40.7128
                      latitude: -74.0060
                    specialRequirements: []
                    status: "pending"
                    totalPrice: 0.00
                    vendorId: 42
                    customerId: "123e4567-e89b-12d3-a456-426614174000"
                    orderTime: "2023-12-07T15:30:00Z"
        '400':
          description: Bad Request - No location present (and no default address for customer) or other input errors (invalid format).
        '401':
          description: Unauthorized - Not a customer user.
        '404':
          description: Not Found - User does not exist.
        '500':
          description: Internal Server Error - An unexpected error occurred on the server.
  /customer/{customerId}/vendors:
    get:
      tags:
        - Customer
      summary: Get list of vendors
      description:  >
        Get a list of vendors within a specified radius around the customer's current location or default address. 
        The radius is predefined - 5 kilometers). The current location is retrieved from the users microservice. If there is no current location, we check if the user has a home address and use that. If there is no home address either, we send an error response. Additionaly, searching and filtering restaurants through query parameters is possible.
      operationId: getVendors
      parameters:
        - in: path
          name: customerId
          required: true
          schema:
            type: string
            format: uuid
          examples:
            customerIdExample:
              value: "123e4567-e89b-12d3-a456-426614174000"    
        - in: query
          name: name
          required: false
          schema:
            type: string      
          examples:
            nameFilterExample:
              value: "McDonalds"  
        - in: query
          name: minAvgPrice
          required: false
          schema:
            type: integer
          examples:
            minPriceExample:
              value: 0  
        - in: query
          name: maxAvgPrice
          required: false
          schema:
            type: integer   
          examples:
            maxPriceExample:
              value: 999    
      responses:
        '200':
          description: List of vendors.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Vendor'
              examples:
                vendorListExample:
                  summary: Example response of a list of vendors
                  value:
                    - ID: "987e4567-e89b-12d3-a456-423144174001"
                      name: "NY pizza"
                      location:
                        houseNumber: 456
                        zip: "12346"
                        Longitude: 40.7130
                        Latitude: -74.0070
                      
        '400':
          description: Bad Request - Invalid request parameters.
        '401':
          description: Unauthorized - Not a customer user.
        '404':
          description: Not Found - User does not exist.
        '500':
          description: Internal Server Error - An unexpected error occurred on the server.
...
  /customer/{customerId}/order/{orderId}/vendor:
    get:
      tags:
        - Customer
      summary: Get all dishes of the selected vendor for the order.
      description: >
        Get a list of all of the dishes of the vendor associated with the order (the selected vendor) as a customer. Only show dishes that don't include the customer's allergies.
      operationId: getVendorDishes
      parameters: 
        - in: path
          name: customerId
          required: true
          schema:
            type: string
            format: uuid
          examples: 
            customerIdExample:
              value:  "123e4567-e89b-12d3-a456-426614174000"    
        - in: path
          name: orderId
          required: true
          schema:
            type: string
            format: uuid
          examples:
            orderIdExample:
              value:  "987e4567-e89b-12d3-a456-423144174001"
      responses:
        '200':
          description: A list of dishes offered by the vendor.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Dish'
              examples:
                example1:
                  summary: Example list of dishes
                  value:
                    - ID: "ccf1f2d7-8f7f-4044-9a30-16790a432977"
                      name: "Margherita Pizza"
                      imageLink: "http://example.com/images/margherita.jpg"
                      price: 12.50
                      allergens: ["Gluten", "Dairy"]
                      ingredients: ["Flour", "Tomatoes", "Mozzarella", "Basil", "Olive Oil"]
                      description: "Best pizza in Delft!"
                    - ID: "0763eb65-564d-4655-bb16-1f0ab46d0c83"
                      name: "Extra Vegan Vegan Salad"
                      imageLink: "http://example.com/images/vegan-salad.jpg"
                      price: 99.99
                      allergens: ["Meat"]
                      ingredients: ["Lettuce", "Extra Lettuce"]
                      description: "A vegan salad"
        '400':
          description: Bad Request - Invalid request parameters.              
        '401':
          description: Unauthorized - User is not a customer/order does not belong to user.
        '404':
          description: User or order not found        
        '500':
          description: Internal Server Error - An unexpected error occurred on the server.
...
  /vendor/{vendorId}/dish/{dishId}:
    get:
      tags:
        - Vendor
      summary: Get details of a specific dish
      description: Get detailed information about a specific dish offered by the vendor.
      operationId: getDish
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: dishId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Detailed information about the specified dish.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
              examples:
                example1:
                  summary: Example dish details
                  value:
                    ID: "ccf1f2d7-8f7f-4044-9a30-16790a432977"
                    name: "Margherita Pizza"
                    imageLink: "http://example.com/images/margherita.jpg"
                    price: 12.50
                    allergens: ["Gluten", "Dairy"]
                    ingredients: ["Flour", "Tomatoes", "Mozzarella", "Basil", "Olive Oil"]
                    description: "Best pizza in Delft!"
        '400':
          description:  Bad Request - Invalid request parameters.               
        '401':
          description: Unauthorized - User is a vendor/dish does not belong to vendor.
        '404':
          description: Dish or vendor not found.
        '500':
          description: Internal Server Error - An unexpected error occurred on the server.

    delete:
      tags:
        - Vendor
      summary: Delete a dish from the menu
      operationId: removeDishFromMenu
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: dishId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Dish removed successfully from the menu.
        '400':
          description: Bad Request - Dish not deleted.
        '401':
          description: Unauthorized - User is not a vendor/dish does not belong to vendor.
        '404':
          description: Not Found - Dish or vendor not found.
        '500':
          description: Internal Server Error - An unexpected error occurred.        
     
    put:
      tags:
        - Vendor
      summary: Update details of a dish
      operationId: updateDishDetails
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: dishId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Dish'
            examples:
              example1:
                summary: Example dish update
                value:
                  name: "Updated Margherita Pizza"
                  imageLink: "http://example.com/new-images/margherita.jpg"
                  price: 13.00
                  allergens: ["Gluten", "Dairy"]
                  ingredients: ["Flour", "Tomatoes", "Mozzarella Cheese", "Basil", "Olive Oil"]
                  description: "Now the best in Hague too!"  
      responses:
        '200':
          description: Dish details updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
              examples:
                exampleResponse:
                  summary: Example response after updating a dish
                  value:
                    ID: "ccf1f2d7-8f7f-4044-9a30-16790a432977"
                    name: "Updated Margherita Pizza"
                    imageLink: "http://example.com/new-images/margherita.jpg"
                    price: 13.00
                    allergens: ["Gluten", "Dairy"]
                    ingredients: ["Flour", "Tomatoes", "Mozzarella Cheese", "Basil", "Olive Oil"]
                    description: "Now the best in Hague too!"
        '400':
          description: Bad Request - Incorrect dish details format or missing required fields.
        '401':
          description: Unauthorized - User is not a vendor/dish does not belong to vendor.
        '404':
          description: Not Found - Dish or vendor not found.
        '500':
          description: Internal Server Error - An unexpected error occurred.
...
components:
  schemas:
    Order:
      type: object
      x-class-extra-annotation: '@javax.persistence.Entity @javax.persistence.Table(name = "customer_order")'
      properties:
        ID:
          type: string
          format: uuid
          x-field-extra-annotation: '@javax.persistence.Id @javax.persistence.GeneratedValue(generator = "UUID") @org.hibernate.annotations.GenericGenerator(name = "UUID", strategy = "org.hibernate.id.UUIDGenerator")'
        dishes:
          type: array
          items:
            $ref: '#/components/schemas/OrderedDish'
          x-field-extra-annotation: '@javax.persistence.OneToMany(cascade=javax.persistence.CascadeType.ALL)'
        location:
          $ref: '#/components/schemas/Address'
        specialRequirements:
          type: string
        status:
          $ref: '#/components/schemas/Status'
        totalPrice:
          type: number
          format: double
        orderTime:
          type: string
          format: date-time
          description: The timestamp when the order was created.  
        vendorId:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
    Address:
      x-class-extra-annotation: '@javax.persistence.Embeddable'
      type: object
      properties:
        houseNumber:
          type: integer
        zip:
          type: string
        longitude:
          type: number
          format: double
        latitude:
          type: number
          format: double
    Vendor:
      type: object
      properties:
        ID:
          type: string
          format: uuid
        name:
          type: string
        location:
          $ref: '#/components/schemas/Address'
    Dish:
      type: object
      x-class-extra-annotation: '@javax.persistence.Entity @javax.persistence.Table(name = "dish")'
      properties:
        ID:
          type: string
          format: uuid
          x-field-extra-annotation: '@javax.persistence.Id @javax.persistence.GeneratedValue(generator = "UUID") @org.hibernate.annotations.GenericGenerator(name = "UUID", strategy = "org.hibernate.id.UUIDGenerator")'
        name:
          type: string
        imageLink:
          type: string
        price:
          type: number
          format: double
        isDeleted:
          type: boolean
          default: false
        allergens:
          type: array
          items:
            type: string
          x-field-extra-annotation: '@javax.persistence.ElementCollection'
        ingredients:
          type: array
          items:
            type: string
          x-field-extra-annotation: '@javax.persistence.ElementCollection'
        description:
          type: string
        vendorId:
          type: string
          format: uuid
    OrderedDish:
      type: object
      x-class-extra-annotation: '@javax.persistence.Entity @javax.persistence.Table(name = "ordered_dish")'
      properties:
        id:
          type: string
          format: uuid
          x-field-extra-annotation: '@javax.persistence.Id @javax.persistence.GeneratedValue(generator = "UUID") @org.hibernate.annotations.GenericGenerator(name = "UUID", strategy = "org.hibernate.id.UUIDGenerator")'
        dish:
          allOf:
            - $ref: '#/components/schemas/Dish'
          x-field-extra-annotation: '@javax.persistence.ManyToOne'
        quantity:
          type: integer
          default: 0
    Status:
      type: string
      enum:
        - pending
        - accepted
        - rejected
        - preparing
        - givenToCourier
        - onTransit
        - delivered
...